[Feat] 1차 파일 변환
